# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CARMA View API
  description: The core's API for the UI and the MCP server (docs/carma-spec.md, «Контракт 4»). Symbol IDs contain spaces, '/' and '#', so they always travel as the query parameter id.
  version: '0.3'
paths:
  /api/v0/status:
    get:
      summary: Contract versions, fact statistics, platform, model counts
      operationId: status_api_v0_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/view:
    get:
      summary: Nodes, edges, issues and pinned positions of one level
      operationId: view_api_v0_view_get
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            default: root
            title: Scope
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/View'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/components:
    get:
      summary: Component tree with symbol counts and issues
      operationId: components_api_v0_components_get
      parameters:
        - name: parent
          in: query
          required: false
          schema:
            type: string
            default: root
            title: Parent
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentTree'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/components/{component_id}:
    get:
      summary: Component model, computed issues, interface symbols, incoming and outgoing edges
      operationId: component_api_v0_components__component_id__get
      parameters:
        - name: component_id
          in: path
          required: true
          schema:
            type: string
            title: Component Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentCard'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/symbols/search:
    get:
      summary: Symbols by name
      operationId: search_api_v0_symbols_search_get
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            title: Q
        - name: kind
          in: query
          required: false
          schema:
            anyOf:
              - type: array
                items:
                  type: string
              - type: 'null'
            description: repeat for several kinds
            title: Kind
          description: repeat for several kinds
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 500
            minimum: 1
            default: 50
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolList'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/symbols:
    get:
      summary: A symbol, its definition, component and editor link
      operationId: symbol_api_v0_symbols_get
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolCard'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/symbols/callers:
    get:
      summary: Who calls the symbol
      operationId: callers_api_v0_symbols_callers_get
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            title: Id
        - name: depth
          in: query
          required: false
          schema:
            type: integer
            maximum: 10
            minimum: 1
            default: 1
            title: Depth
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 1000
            minimum: 1
            default: 200
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallList'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/symbols/callees:
    get:
      summary: What the symbol calls
      operationId: callees_api_v0_symbols_callees_get
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            title: Id
        - name: depth
          in: query
          required: false
          schema:
            type: integer
            maximum: 10
            minimum: 1
            default: 1
            title: Depth
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 1000
            minimum: 1
            default: 200
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallList'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/paths:
    get:
      summary: Shortest call chains between two symbols
      operationId: paths_api_v0_paths_get
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            title: From
        - name: to
          in: query
          required: true
          schema:
            type: string
            title: To
        - name: max_depth
          in: query
          required: false
          schema:
            type: integer
            maximum: 12
            minimum: 1
            default: 6
            title: Max Depth
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 50
            minimum: 1
            default: 5
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PathList'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/edges/samples:
    get:
      summary: The refs an edge of a level is made of
      operationId: samples_api_v0_edges_samples_get
      parameters:
        - name: src
          in: query
          required: true
          schema:
            type: string
            title: Src
        - name: dst
          in: query
          required: true
          schema:
            type: string
            title: Dst
        - name: scope
          in: query
          required: false
          schema:
            type: string
            default: root
            title: Scope
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 500
            minimum: 1
            default: 20
            title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefList'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/layout/{scope}:
    put:
      summary: Save the pinned positions of one level
      operationId: layout_api_v0_layout__scope__put
      parameters:
        - name: scope
          in: path
          required: true
          schema:
            type: string
            title: Scope
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LayoutBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LayoutResult'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/events:
    get:
      summary: 'Server-sent events: facts_changed, model_changed, highlight_changed, reindex_progress'
      operationId: events_api_v0_events_get
      responses:
        '200':
          description: an event stream
          content:
            text/event-stream: {}
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/workspace:
    get:
      summary: Renderer-independent viewing state
      operationId: workspace_api_v0_workspace_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Save the last level, camera and filters
      operationId: write_workspace_api_v0_workspace_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Workspace'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v0/layout-cache/{key}:
    get:
      summary: Derived node positions for this project and graph
      operationId: layout_cache_api_v0_layout_cache__key__get
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            title: Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CachedLayout'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Write Layout Cache
      operationId: write_layout_cache_api_v0_layout_cache__key__put
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
            title: Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CachedLayout'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CachedLayout'
        '404':
          description: unknown ID or scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ComponentTree:
      properties:
        parent:
          type: string
          title: Parent
        components:
          items:
            $ref: '#/components/schemas/TreeNode'
          type: array
          title: Components
        unassigned:
          type: integer
          title: Unassigned
          description: symbols in no component, shown as the _unassigned node at the top level
      additionalProperties: false
      type: object
      required:
        - parent
        - components
        - unassigned
      title: ComponentTree
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      additionalProperties: false
      type: object
      required:
        - error
      title: ErrorResponse
...
    Status:
      properties:
        version:
          type: string
          title: Version
          description: carma package version
        contracts:
          $ref: '#/components/schemas/Contracts'
        project_root:
          type: string
          title: Project Root
        project_name:
          type: string
          title: Project Name
        facts:
          $ref: '#/components/schemas/FactsStatus'
        components:
          type: integer
          title: Components
        invalid_files:
          items:
            type: string
          type: array
          title: Invalid Files
        issues:
          type: integer
          title: Issues
        highlight:
          type: boolean
          title: Highlight
      additionalProperties: false
      type: object
      required:
        - version
        - contracts
        - project_root
        - project_name
        - facts
        - components
        - invalid_files
        - issues
        - highlight
      title: Status
...
    SymbolList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/SymbolBrief'
          type: array
          title: Items
        more:
          type: boolean
          title: More
          description: true when the list was cut at limit
      additionalProperties: false
      type: object
      required:
        - items
        - more
      title: SymbolList
...